+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Add `salmon.util` functions `full-name`, `tags`, and `resource`.
   These are useful when building CloudFormation templates.
 - Add `:tags` option to stack for defining stack-level tags.
+- Add `:tags` and `:tags-raw` maps to stack and stack-properties
+  instances.
 ## v0.12.0 (2023-11-01)
 
 - Add `:client` and `:region` options to stack and stack-properties